COM IA FUNCIONANDO

+++ backend/src/controllers/QuickAnswerController.ts
@@ -32,6 +32,10 @@ export const index = async (req: Request, res: Response): Promise<Response> => {
 };
 
 export const store = async (req: Request, res: Response): Promise<Response> => {
+  if (req.user.profile !== "admin") {
+    throw new AppError("ERR_NO_PERMISSION", 403);
+  }
+
   const newQuickAnswer: QuickAnswerData = req.body;
 
   const QuickAnswerSchema = Yup.object().shape({
@@ -70,6 +74,10 @@ export const update = async (
   req: Request,
   res: Response
 ): Promise<Response> => {
+  if (req.user.profile !== "admin") {
+    throw new AppError("ERR_NO_PERMISSION", 403);
+  }
+
   const quickAnswerData: QuickAnswerData = req.body;
 
   const schema = Yup.object().shape({
@@ -103,6 +111,10 @@ export const remove = async (
   req: Request,
   res: Response
 ): Promise<Response> => {
+  if (req.user.profile !== "admin") {
+    throw new AppError("ERR_NO_PERMISSION", 403);
+  }
+
   const { quickAnswerId } = req.params;
 
   await DeleteQuickAnswerService(quickAnswerId);
